Added scissor test to the render states.
Former-commit-id: 0e51f5c6d4a1f971a740bc9902eaecbcad7950f6 [formerly 2d1f2786b22e949a09fc60e1541ae415e19746e3]
Former-commit-id: e8776d5aaada354a8bc5ec82cd227f42aa75ee90 [formerly ffa9853cb947ccdfe0a8e4f3e5968895ab1e7eb6]
Former-commit-id: 6b98507bd0901ec983b37b7a089d98518d5adb14

+++ examples/line-geometry/src/Main.cpp
@@ -120,7 +120,7 @@ addStar(Node::Ptr root, file::AssetLibrary::Ptr assets, std::vector<Star>& stars
 				->set("material.lineThickness",	1.0f + 3.0f * (rand() / (float)RAND_MAX)),
 			assets->effect("line")
 		))
-		->addComponent(Transform::create(Matrix4x4::create()->appendRotationZ(2.0f * PI * rand() / (float)RAND_MAX)));
+		->addComponent(Transform::create(Matrix4x4::create()->appendRotationZ(2.0f * (float)PI * rand() / (float)RAND_MAX)));
 	
 	stars.push_back(Star());
 	
@@ -142,8 +142,8 @@ createStarLineGeometry(unsigned int numStarBranches,
 	const unsigned int	numBranches = std::max((unsigned int)3, numStarBranches);
 	const float			innerRadius	= std::min(inRadius, outRadius);
 	const float			outerRadius	= std::max(inRadius, outRadius);
-	const float			cStep		= cosf(PI / (float)numBranches);
-	const float			sStep		= sinf(PI / (float)numBranches);
+	const float			cStep		= cosf((float)PI / (float)numBranches);
+	const float			sStep		= sinf((float)PI / (float)numBranches);
 
 	LineGeometry::Ptr	lines		= LineGeometry::create(context);
